parallel-ecommerce-recommender/requirements_fixed.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=5.0.0
scikit-learn>=1.5.0
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
faker>=20.0.0